Refactor code
- clean up page component by creating separate PublishArticleForm component
- modify errors and messages to be in Macedonian
- make page component a server component
- modify eslint rules to avoid false positive for inputs
- remove duplicate label in TagList component ("Tags:")
